% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddingViz.R
\name{analyze_embedding_separation}
\alias{analyze_embedding_separation}
\title{Analyze the characteristics of embedding clusters}
\usage{
analyze_embedding_separation(
  embedding_matrix,
  metadata_df,
  metadata_cols = NULL
)
}
\description{
Analyze the characteristics of embedding clusters
}
