% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureExtraction.R
\name{extract_population_features}
\alias{extract_population_features}
\title{Extract population-level features from all nuclei}
\usage{
extract_population_features(nuclei_list, verbose = FALSE)
}
\arguments{
\item{nuclei_list}{List of nuclei}

\item{verbose}{Progress messages flag}
}
\value{
List of population statistics
}
\description{
Extract population-level features from all nuclei
}
