% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_extraction.R
\name{extract_spatial_features}
\alias{extract_spatial_features}
\title{Extract spatial pattern features from nucleus centroids}
\usage{
extract_spatial_features(nuclei_list, image_dims, verbose = FALSE)
}
\arguments{
\item{nuclei_list}{List of nuclei}

\item{image_dims}{Image dimensions c(width, height)}

\item{verbose}{Progress messages flag}
}
\value{
List of spatial analysis results
}
\description{
Extract spatial pattern features from nucleus centroids
}
\details{
Performs spatial point pattern analysis including:
         - Nuclear density calculations
         - Nearest neighbor analysis
         - Clustering analysis using Ripley's K-function
         - Spatial regularity metrics
}
