% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_extraction.R
\name{fit_ellipse_to_contour}
\alias{fit_ellipse_to_contour}
\title{Fit ellipse to contour and extract parameters}
\usage{
fit_ellipse_to_contour(contour)
}
\arguments{
\item{contour}{Matrix of x,y boundary coordinates}
}
\value{
List with ellipse parameters
}
\description{
Fit ellipse to contour and extract parameters
}
