% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEmbedding.R
\name{getEmbedding}
\alias{getEmbedding}
\title{Select and expand Prov-GigaPath embeddings}
\usage{
getEmbedding(pg, layer = NULL)
}
\arguments{
\item{pg}{A data frame. Prov-GigaPath output with the dimension 1x15.}

\item{layer}{The layer of embedding. Available 0 to 13. If it is set to 
`NULL`, all the layers will be returned as 'samples x feature_dimensions'
matrix.}
}
\value{
A data frame (1 x 768) of a chosen embedding layer
}
\description{
Select and expand Prov-GigaPath embeddings
}
