% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importImgFeatures.R
\name{importImgFeatures}
\alias{importImgFeatures}
\title{Import nucleu features for TCGA samples as SpatialExperiment}
\usage{
importImgFeatures(fnames, subsample_rate = 10, seed = NULL)
}
\arguments{
\item{subsample_rate}{A numeric value >= 1. `1/subsample_rate` amount of 
cells will be collected. If you want to import all the cells, set this as `1`.}

\item{x}{A character vector containing file paths to image feature files.}
}
\value{
An ExperimentList object containing SpatialExperiment objects (one per slide/sample)
}
\description{
This function imports HoVerNet outputs (JSON) or Squidpy outputs (h5ad) as R list.
}
