% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonToSfe.R
\name{jsonToSfe}
\alias{jsonToSfe}
\title{Import HoVerNet output (JSON) as SpatialFeatureExperiment}
\usage{
jsonToSfe(fname, subsample_rate = subsample_rate, seed = seed)
}
\arguments{
\item{fname}{A character(1). Containing a file path to an image feature file.
This function expects that file names include the TCGA patient ID in a `TCGA-\\d{2}-\\d{4}` pattern.}

\item{subsample_rate}{A numeric value >= 1. `1/subsample_rate` amount of cells will be collected.
If you want to import all the cells, set this as `1`.}

\item{seed}{Integer or NULL. Random seed for reproducible results. 
If NULL (default), no seed is set and results will be random.}
}
\value{
A SpatialFeatureExperiment object
}
\description{
Import HoVerNet output (JSON) as SpatialFeatureExperiment
}
